和大家工作进度一致

+++ static/js/login.js
@@ -42,9 +42,22 @@ var resigerVue = new Vue({
                     }
                     else{
                         console.log(response.body);
-                        alert('登陆成功');
-                        sessionStorage.setItem('login',toJSON(response.body));
-                        window.location.href = "/static/html/system.html";
+                        if (response.body.data.confirmed == false) {
+                        	alert("邮箱未验证");
+                        	sessionStorage.setItem('login',toJSON(response.body));
+                        	window.location.href = "/static/html/unconfirmed.html";
+                        } else {
+                        	if (response.body.data.role_name == "管理员") {
+                        		alert('登陆成功');
+		                        sessionStorage.setItem('login',toJSON(response.body));
+		                        window.location.href = "/static/html/system.html";
+                        	} else{
+                        		alert('登陆成功');
+		                        sessionStorage.setItem('login',toJSON(response.body));
+		                        window.location.href = "/static/html/system_for_user.html";
+                        	}
+                        }
+                        
                     }
                 });
         }
